Simplify file hierarchy
