+++ packages/changed-elements-react/src/widgets/comparisonJobWidget/components/VersionCompareSelectDialogV2.scss
