+++ packages/changed-elements-react/src/widgets/comparisonJobWidget/versionCompareToasts.ts
@@ -9,11 +9,11 @@ import { toaster } from "@itwin/itwinui-react";
 
 import type {
   ComparisonJobCompleted, IComparisonJobClient,
-} from "../../../clients/IComparisonJobClient";
-import type { IModelsClient, NamedVersion } from "../../../clients/iModelsClient";
-import type { ComparisonJobUpdateType } from "../components/VersionCompareDialogProvider";
-import type { JobAndNamedVersions } from "../models/ComparisonJobModels";
-import { runManagerStartComparisonV2 } from "./versionCompareV2WidgetUtils";
+} from "../../clients/IComparisonJobClient";
+import type { IModelsClient, NamedVersion } from "../../clients/iModelsClient";
+import type { ComparisonJobUpdateType } from "./components/VersionCompareDialogProvider";
+import type { JobAndNamedVersions } from "./NamedVersions.js";
+import { runManagerStartComparisonV2 } from "./versionCompareV2WidgetUtils.js";
 
 /** Toast Comparison Job Processing.
  * Outputs toast message following the pattern:
